+++ packages/movesense_flutter/CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+Initial release supporting:
+
+* scan for Movensense devices (and stop scanning again)
+* connect and disconnect to a device
+* get device and state information
+* get a stream of the following data from a device:
+  * heart rate
+  * ECG
+  * IMU
+  * temperature
+  * state events (movement, battery, connectors, double tap, tap, free fall)